BAU: Optimize and cleanup some tests
With an introduction of parallel testing we can optimize the runtime even more by
optimizing some of the tests and spreading the tests across features.
Instead of having a couple chunk feature test, it's better to break them
down to multiple and smaller bits.
I also removed some of the redundant tests which are being tested somewhere else.