BAU_run_selected_tests
Parallel cucumber tests mean that if a test fails and needs to be run again the whole feature is the smallest unit that can be selected and requires the pre-commit script to be edited.

Add script to take arguments of the form feature/some_test.feature:9 to run only the test starting at line 9 of some_test.feature.  Multiple tests of that format can be specified if more than one test is failing.  An optional argument of --show-browser can be specified.