Add MSA backward compatibility tests
This adds tests for MSA backward compatibility.  They reproduce the
old ida-hub-acceptance-tests/msa-backward-compatibility-test.sh which
ran
MsaBackwardsCompatibilityTest.userSignInSuccessfullyAndCompletesNINOCycle3Page.

Basically it's a simple login flow with cycle 3.

The purpose of these tests is that they are run against a separate
test-rp/msa instance, which are running an old but supported version
of the MSA.  These MSAs are already deployed to PaaS and have
pipelines to deploy them.

It's not ideal to hardcode the MSA URL in the feature but I couldn't
think of a better way to achieve this.  I'm open to suggestions.

I have run these locally and they pass :+1: so we are still backwards
compatible, hooray!