BAU: Fix Belgian connector-node smoke tests
The Belgians have added a new banner to their landing page which
includes the work "problems". This matches with our error checking step
which has a regex for "problem".

This commit adds a word boundary to the regex so we don't match the
plural. This is a bit fragile, but it at least preserves some of the
functionality.