BAU: Remove use of Selenium grid
We've found recently that the Selenium grid running in ECS has been very
unreliable and has prevented releases from progressing though the hub
pipeline. Acceptance tests would often timeout, amongst other errors.
The grid show's high use of memory and CPU which could be a contributing
factor. Attempts to diagnose the reason for this flakiness have been
fruitless.

This commit switches to using a local firefox driver. This seems to work
reliably, and means if we can stop using the grid we have one less thing
to maintain/worry about.

The logic is slightly re-ordered in the `env.rb` file so that the case
when we're running in Docker is the special case that uses a Selenium
grid (see the docker-compose file).