[HUB-841] deploy-stubs requires loaded images
The run-acceptance-tests job previously did not load the images due to
it being a cdind job rather than we're actually building something.

In the verify-infrastructure-config pipeline for deploy-stubs I have
added the images to the job as oci images that we can then docker load
here. The configuration is rather disconnected and will probably need
some documentation.

Signed-off-by: Chris Mills <7100370+chriscoffee@users.noreply.github.com>